+++ src/components/common/Search.jsx
@@ -18,14 +18,14 @@ const StyledSearch = styled.div`
   }
 `;
 
-const Search = () => {
+const Search = ({ placeholder }) => {
   const onChange = () => {};
   return (
     <StyledSearch>
       <h3>Search</h3>
       <Input
         type="text"
-        placeholder="Look for a specific excercise"
+        placeholder={"Look for a specific excercise"}
         id="search"
         name="search"
         variant="filled"
