+++ src/components/exercise/ExerciseList.jsx
@@ -4,19 +4,15 @@ import Excercises from "./Exercises";
 import { ExercisesStyle } from "./ExerciseStyle";
 import Search from "../common/Search";
 
-function Exercises() {
+const Exercises = () => {
   return (
     <div>
       <ExercisesStyle>
-        <Search />
+        <Search placeholder="Look for a specific excercise" />
         <Excercises />
-        {/* <div className="recommendations-container">
-          <p>See recommended exercises</p>
-          <p>These exercises were suggested based on your current progress</p>
-        </div> */}
       </ExercisesStyle>
     </div>
   );
-}
+};
 
 export default Exercises;
