+++ src/components/workouts/WorkoutStyle.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const WorkoutsStyle = styled.div`
+  width: 80vw;
+  margin: 0 auto;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  padding: 2rem 0;
+`;
+
+export const WorkoutStyle = styled.div`
+  display: flex;
+`;
